solidity linting

+++ contracts/Yielder.sol
@@ -11,15 +11,21 @@ import "./creamfi-helper/ComptrollerInterface.sol";
 
 contract Yielder is ContractWithFlashLoan, Ownable {
     using SafeMath for uint256;
-    
+
     address public constant ETHER = 0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE;
 
     constructor(address _aaveLPProvider)
-        public payable
+        public
+        payable
         ContractWithFlashLoan(_aaveLPProvider)
     {}
 
-    function start(address cTokenAddr, uint256 flashLoanAmount, bool isCEther, bool windYield) public payable onlyOwner {
+    function start(
+        address cTokenAddr,
+        uint256 flashLoanAmount,
+        bool isCEther,
+        bool windYield
+    ) public payable onlyOwner {
         address loanToken;
         if (isCEther) {
             loanToken = ETHER;
@@ -28,7 +34,12 @@ contract Yielder is ContractWithFlashLoan, Ownable {
             loanToken = CErc20Interface(cTokenAddr).underlying();
         }
 
-        bytes memory params = abi.encode(msg.sender, cTokenAddr, isCEther, windYield);
+        bytes memory params = abi.encode(
+            msg.sender,
+            cTokenAddr,
+            isCEther,
+            windYield
+        );
 
         initateFlashLoan(address(this), loanToken, flashLoanAmount, params);
     }
@@ -44,7 +55,10 @@ contract Yielder is ContractWithFlashLoan, Ownable {
         bool isCEther;
         bool windYield;
 
-        (messageSender, cTokenAddr, isCEther, windYield) = abi.decode(params, (address, address, bool, bool));
+        (messageSender, cTokenAddr, isCEther, windYield) = abi.decode(
+            params,
+            (address, address, bool, bool)
+        );
         require(owner() == messageSender, "caller is not the owner");
 
         if (windYield) {
@@ -59,14 +73,24 @@ contract Yielder is ContractWithFlashLoan, Ownable {
             return;
         }
 
-        uint loanRepayAmount = amount.add(fees);
-        uint loanedTokenBalOfThis = IERC20(loanedToken).balanceOf(address(this));
+        uint256 loanRepayAmount = amount.add(fees);
+        uint256 loanedTokenBalOfThis = IERC20(loanedToken).balanceOf(
+            address(this)
+        );
         if (loanedTokenBalOfThis < loanRepayAmount) {
-            IERC20(loanedToken).transferFrom(messageSender, address(this), loanRepayAmount - loanedTokenBalOfThis);
+            IERC20(loanedToken).transferFrom(
+                messageSender,
+                address(this),
+                loanRepayAmount - loanedTokenBalOfThis
+            );
         }
     }
 
-    function supplyToCream(address cTokenAddr, uint amount, bool isCEther) public payable onlyOwner returns (bool) {
+    function supplyToCream(
+        address cTokenAddr,
+        uint256 amount,
+        bool isCEther
+    ) public payable onlyOwner returns (bool) {
         if (isCEther) {
             CEtherInterface(cTokenAddr).mint.value(amount)();
         } else {
@@ -74,11 +98,15 @@ contract Yielder is ContractWithFlashLoan, Ownable {
             checkBalThenTransferFrom(underlying, msg.sender, amount);
             checkThenErc20Approve(underlying, cTokenAddr, amount);
             cTokenMint(cTokenAddr, amount);
-        }    
+        }
         return true;
     }
 
-    function repayBorrowedFromCream(address cTokenAddr, uint amount, bool isCEther) public payable onlyOwner returns (bool) {
+    function repayBorrowedFromCream(
+        address cTokenAddr,
+        uint256 amount,
+        bool isCEther
+    ) public payable onlyOwner returns (bool) {
         if (isCEther) {
             CEtherInterface(cTokenAddr).repayBorrow.value(amount)();
         } else {
@@ -86,38 +114,58 @@ contract Yielder is ContractWithFlashLoan, Ownable {
             checkBalThenTransferFrom(underlying, msg.sender, amount);
             checkThenErc20Approve(underlying, cTokenAddr, amount);
             cTokenRepayBorrow(cTokenAddr, amount);
-        }    
+        }
         return true;
     }
 
-    function withdrawFromCream(address cTokenAddr, uint amount) public onlyOwner returns (bool) {
+    function withdrawFromCream(address cTokenAddr, uint256 amount)
+        public
+        onlyOwner
+        returns (bool)
+    {
         return cTokenRedeemUnderlying(cTokenAddr, amount);
     }
 
-    function borrowFromCream(address cTokenAddr, uint amount) public onlyOwner returns (bool) {
+    function borrowFromCream(address cTokenAddr, uint256 amount)
+        public
+        onlyOwner
+        returns (bool)
+    {
         return cTokenBorrow(cTokenAddr, amount);
     }
 
-    function cTokenMint(address cToken, uint mintAmount) internal returns (bool) {
-        uint err = CErc20Interface(cToken).mint(mintAmount);
+    function cTokenMint(address cToken, uint256 mintAmount)
+        internal
+        returns (bool)
+    {
+        uint256 err = CErc20Interface(cToken).mint(mintAmount);
         require(err == 0, "cToken mint failed");
         return true;
     }
 
-    function cTokenRedeemUnderlying(address cToken, uint redeemAmount) internal returns (bool) {
-        uint err = CErc20Interface(cToken).redeemUnderlying(redeemAmount);
+    function cTokenRedeemUnderlying(address cToken, uint256 redeemAmount)
+        internal
+        returns (bool)
+    {
+        uint256 err = CErc20Interface(cToken).redeemUnderlying(redeemAmount);
         require(err == 0, "cToken redeem failed");
         return true;
     }
 
-    function cTokenBorrow(address cToken, uint borrowAmount) internal returns (bool) {
-        uint err = CErc20Interface(cToken).borrow(borrowAmount);
+    function cTokenBorrow(address cToken, uint256 borrowAmount)
+        internal
+        returns (bool)
+    {
+        uint256 err = CErc20Interface(cToken).borrow(borrowAmount);
         require(err == 0, "cToken borrow failed");
         return true;
     }
 
-    function cTokenRepayBorrow(address cToken, uint repayAmount) internal returns (bool) {
-        uint err = CErc20Interface(cToken).repayBorrow(repayAmount);
+    function cTokenRepayBorrow(address cToken, uint256 repayAmount)
+        internal
+        returns (bool)
+    {
+        uint256 err = CErc20Interface(cToken).repayBorrow(repayAmount);
         require(err == 0, "cToken repay failed");
         return true;
     }
@@ -126,39 +174,66 @@ contract Yielder is ContractWithFlashLoan, Ownable {
         ComptrollerInterface(comptroller).claimComp(address(this));
     }
 
-    function claimAndTransferCream(address comptrollerAddr, address receiver) public onlyOwner {
-
-        ComptrollerInterface comptroller = ComptrollerInterface(comptrollerAddr);
+    function claimAndTransferCream(address comptrollerAddr, address receiver)
+        public
+        onlyOwner
+    {
+        ComptrollerInterface comptroller = ComptrollerInterface(
+            comptrollerAddr
+        );
         comptroller.claimComp(address(this));
 
         IERC20 compToken = IERC20(comptroller.getCompAddress());
-        uint totalCompBalance = compToken.balanceOf(address(this));
+        uint256 totalCompBalance = compToken.balanceOf(address(this));
 
-        require(compToken.transfer(receiver, totalCompBalance), "cream transfer failed");
+        require(
+            compToken.transfer(receiver, totalCompBalance),
+            "cream transfer failed"
+        );
     }
 
-    function claimAndTransferCreamForCToken(address comptrollerAddr, address[] memory cTokens, address receiver) public onlyOwner {
-        ComptrollerInterface comptroller = ComptrollerInterface(comptrollerAddr);
+    function claimAndTransferCreamForCToken(
+        address comptrollerAddr,
+        address[] memory cTokens,
+        address receiver
+    ) public onlyOwner {
+        ComptrollerInterface comptroller = ComptrollerInterface(
+            comptrollerAddr
+        );
         comptroller.claimComp(address(this), cTokens);
 
         IERC20 compToken = IERC20(comptroller.getCompAddress());
-        uint totalCompBalance = compToken.balanceOf(address(this));
+        uint256 totalCompBalance = compToken.balanceOf(address(this));
 
-        require(compToken.transfer(receiver, totalCompBalance), "cream transfer failed");
+        require(
+            compToken.transfer(receiver, totalCompBalance),
+            "cream transfer failed"
+        );
     }
 
-    function checkBalThenTransferFrom(address tokenAddress, address user, uint amount) internal returns (bool) {
-        uint balOfThis = IERC20(tokenAddress).balanceOf(address(this));
+    function checkBalThenTransferFrom(
+        address tokenAddress,
+        address user,
+        uint256 amount
+    ) internal returns (bool) {
+        uint256 balOfThis = IERC20(tokenAddress).balanceOf(address(this));
         if (balOfThis < amount) {
             IERC20(tokenAddress).transferFrom(user, address(this), amount);
         }
         return true;
     }
 
-    function checkThenErc20Approve(address tokenAddress, address approveTo, uint amount) internal returns (bool) {
-        uint allowance = IERC20(tokenAddress).allowance(address(this), approveTo);
+    function checkThenErc20Approve(
+        address tokenAddress,
+        address approveTo,
+        uint256 amount
+    ) internal returns (bool) {
+        uint256 allowance = IERC20(tokenAddress).allowance(
+            address(this),
+            approveTo
+        );
         if (allowance < amount) {
-            IERC20(tokenAddress).approve(approveTo, uint(-1));
+            IERC20(tokenAddress).approve(approveTo, uint256(-1));
         }
         return true;
     }
@@ -195,11 +270,11 @@ contract Yielder is ContractWithFlashLoan, Ownable {
         return true;
     }
 
-    function tokenBalance(address token) public view returns (uint balance) {
+    function tokenBalance(address token) public view returns (uint256 balance) {
         return IERC20(token).balanceOf(address(this));
     }
 
-    function ethBalance() public view returns (uint balance) {
+    function ethBalance() public view returns (uint256 balance) {
         return address(this).balance;
     }
 }
